Removing comment regarding dual license
Fedora Packaging Guidelines document a number of different
dual/multiple licensing scenarios which I didn't read in its
entirety and therefore over-looked the one line that mattered.
Because the contents of the binary package [not the source tarbal]
is all BSD the dual license on the base package is not required.