don't skip metadata checks when using lock-method = auto-detect
Previously we were asking the schema discovery code to unconditionally
skip table metadata when estimate-method != plugin to avoid expensive
calculations when no table filtering was required. This is not safe
to do with lock-method = auto-detect since we have to at least look at
engine metadata to determine what lock method to use.