implement config formatter functionality
Config objects now have a 'formatter' attribute - an object with a
format method which is called with a key/value pair.  This formatter object
is used to then serialize a value to a config-friendly string when formatting
config file output.
This is used by the Configspec validate() method to both deserialize
a config value to a useful python object and provide a useful formatter
to serialize this object back into a parsable config file value.