cleanup log message when a non-managed option is passed to mysqldump
A set of mysqldump options are scanned by holland and some basic
safety checks enforced. e.g. flush-privileges isn't supported on mysql 4.1
and dump-events is not supported on 5.0, etc.  When a user passes an explicit
option via additional-options that the mysqldump plugin does not recognize
we flag this as either an 'unparseable option' - which may cause mysqldump to
fail or as a 'user supplied option' - if it looks valid but holland doesn't
have any special check for it.