handle failing SHOW CREATE VIEW when exclude-invalid-views is enabled
There was a bug in the initial view exclusion code that would raise
an unchecked exception when SHOW CREATE VIEW failed.  This is caused
when retrieving a defintion for a view that references columns that
no longer exist on an underlying table - likely someone ran an
"ALTER TABLE ... DROP COLUMN ..." statement.  This new implementation
will fallback to reconstructing a view from INFORMATION_SCHEMA.VIEWS
if SHOW CREATE VIEW fails.  This will miss a few view options but at
least retrieve the vast majority of the view definition.