fix serious bug where streaming backps redicted to /dev/null
When determining the filename where a backup should be stored,
the xtrabackup plugin only considered values for stream of 'tar'
and 'xbstream', so when using the 1.0.6 default value of
stream = yes, this would generate an empty backup.
This patch ensures we're always comparing the stream value
through determine_stream-method() so we have a consistent
view of the real backup method request, where boolean true
values map to 'tar' and false values map to a non-streaming
method.