ensure config validation method 'force_list' treats empty strings as empty lists
Previously a list accepting option that was set to an empty string would
generate a list with a single empty-string vaue.  This would cause
confusing errors and require additional checks to work around.

This patch ensures force_list will return an empty list of a value is
a false value. so "option = " is equivalent to "option = ,"