always execute FLUSH LOCAL TABLES
This adds FLUSH /*!40101 LOCAL */ TABLES when flush_tables() is
run via the client interface in holland.lib.mysql.  This change
ensures that the FLUSH TABLES command is not written to the binary
log and thus will not replicate down to any replication slaves
that may cause problems, particularly in multi-master environments.