Indented version tags, CSS updates
Indented the version tags where they were being used to describe a particular
command or function in the config and overview documents. Also a few more
CSS updates, largely over <pre> tags - I thought the default didn't make it
obvious enough that it was source code and it was running into other sections.