use TemporaryFile rather than pipe for stderr
stderr can overflow a pipe buffer and lead to deadlocking. With
the configurable compression commands this is more likely
particularly when using options like --verbose.

Notably this change does introduce an additional potential failure
case if there is an io error on the temporary file (ENOSPC, etc.)
but this is likely to cause the backup to fail anyway.