fix bad iteration over rows in mysql50 metadata discovery
This was previously changed as part of a fix for invalid
view detection in MySQL 5.0, where views were never being
detected correctly in order for the invalid view filtering
to work.  Unfortunately this introduced mutating the row
dictionary while iterating over the raw dictionary directly
which results in a "dictionary changed size during iteration"
error.  keys() are not iterated over to avoid that situation.