Revert broken SHOW SLAVE STATUS fix for LP#1220841
The fix for LP#1220841 breaks show_slave_status() in the holland
mysql client 'show_slave_status()' method.  'binary' is not a valid
character set for MySQLdb. This leads to an 'unknown encoding' error
and the backup fails.

Without changing connectors to something other than MySQLdb - where
'raw' binary strings can be had easily - I think LP#1220841 is not
easily fixed.  The original fix for this show slave status bug is
reverted here to address the serious defect it introduced in
holland 1.0.12.