Fix lock-method = auto-detect for MySQL 5.0
There was a regression from the fix for LP#1081261 which caused
storage engine name comparison to be done in lowercase. The engine
attribute was canonicalized in the path for INFORMATION_SCHEMA
(MySQL 5.1+), but not for the "SHOW TABLE STATUS" path (used for
MySQL 5.0 and earlier). This causes the mysqldump option
lock-method=auto-detect to always default to 'non-transactional'
for MySQL 5.0 envionments and use the default mysqldump option
--lock-tables.

Resolves #148