WIP: Fix bugs in holland-mysqldump's "stop-slave" feature
The mysqldump plugin will attempt to restart a slave incorrectly
when stop slave fails.

This change defers injecting the [mysql:replication] section in the
backup.conf config until STOP SLAVE is successful.  This allows
the plugin to detect whether it actually stopped a slave before
attempting to restart it, which was the original intention.

Based off mikegriffin's logging patches around the stop-slave feature.